moving document List to a seperate reducer
